Migrate to Scala 3 syntax: type lambdas
For when we need a generic type with one (value-level) type parameter, but we have a generic type with
two (value-level) type parameters. In those cases we fix one of the type parameters, so we need to
express the resulting new type. E.g. `Either[L, R]` where `L` is fixed to `String` becomes
`[R] =>> Either[String, R]`